Fixed Shipping TransactionEnviron
Shipping was not setting TransactionEnvironment.  It worked locally and
in staging because we defaulted to Sandbox on the proxy side.  Once we
got to production is when the errors began.

We are setting Production based on the keystoneEnvironment variable now
just like we do in the taxationRequestCoordinator and this has solved
the problem.