refactor(SC-1525): Cleaned up Quertle Calls
Found error case with USPTO Site where making too many calls would result in Patent not found page
being returned.  Checking for this now to prevent processing of wrong article as well as cleaned up
Code to only make Full Article Call when we want to process drastically reducing the number of calls
we make.  Also added Error Email that sends when an error message is generated.  Error messages are
user driven, not system.  IE Errors are logged, Error messages may be generated if necessary to let
Mike Know.  In this case it's only generated when we have that bad response from USPTO so he can
manage expectations and we can potentially target a re-run of that article.