feat(smarttrak fd): added row to Financial Dashboard export describing the data source
With the change, if the current segment of the FD is not on the root node, then there will be an
additional line on the excel export.  If a company is slected this will list the name of the
company.  If a market or lower segment is selected, it will list the name of the most specific
market segment.

st 143